Use a separate ws_header_t for sending and receiving
If the code simultaneously sending and receiving a frame, ws->header
ends up being used for both tasks, which means the sending side will
overwrite it and mess up the state needed by the reading side.
Fixed this by adding a separate field ws->send_header for the sending
code to use.